+++ R/predict.R
@@ -123,7 +123,9 @@ function( object,
 #'        into one or more derived quantities.  This can be used to compute area-expanded
 #'        indices for more than one year or category using a single call, and might be
 #'        substantially faster for large models (because it avoids extra model builds
-#'        for each derived quantity)
+#'        for each derived quantity).  Note that the output will have as many rows
+#'        as the highest value of \code{block}, and NAs for rows where \code{block}
+#'        does not contain that integer.
 #' @param type Integer-vector indicating what type of expansion to apply to
 #'        each row of `newdata`, with length of \code{nrow(newdata)}.  
 #' \describe{
@@ -237,7 +239,64 @@ function( object,
 #' estimate if available, and the standard error for the bias-corrected estimator.
 #' Depending upon settings, one or more of these will be \code{NA} values, and the
 #' function can be repeatedly called to get multiple estimators and/or statistics.
-#' 
+#'
+#' @examples
+#' # Settings
+#'  set.seed(123)
+#'  n_sampled = 50
+#'  samples_per_site = 3
+#'  n_extra = 50
+#'
+#' # Simulate densities and lognormal samples
+#'  s_i = rep( seq_len(n_sampled), each = samples_per_site )
+#'  x_s = exp( 1 + rnorm( n_sampled + n_extra ))
+#'  y_i = exp( rnorm(n_sampled * samples_per_site, mean = log(x_s[s_i]), sd = 0.5 ) )
+#'  data = data.frame( y = y_i, s = s_i )
+#'
+#' # Make spatial graph of all sites
+#'  library(igraph)
+#'  unconnected_graph = make_empty_graph( n_sampled + n_extra )
+#'  V(unconnected_graph)$name = as.character(seq_len(n_sampled + n_extra))
+#'
+#' # Fit in tinyVAST
+#'  fit = tinyVAST(
+#'    data = data,
+#'    formula = y ~ 1,
+#'    spatial_domain = unconnected_graph,
+#'    space_column = "s",
+#'    family = lognormal("log"),
+#'    space_term = "response <-> response, spatial_sd"
+#'  )
+#'
+#' # Sample densities
+#'  domain_df = data.frame(s = as.character(seq_len(n_sampled + n_extra)))
+#'  samples = sample_variable(
+#'    fit,
+#'    newdata = domain_df,
+#'    n_samples = 1000,
+#'    sample_fixed = FALSE,
+#'    variable_name = "mu_g"
+#'  )
+#'
+#' # Estimate total
+#' # Note that bias-corrected estimator matches True and sample-based estimator
+#'  total = integrate_output(
+#'    fit,
+#'    newdata = domain_df
+#'  )
+#'
+#' # Compare these options
+#'  c( True = sum(x_s), Sampled = mean(colSums(samples)), total )
+#'
+#' # Calculate effective area occupied from samples
+#'  numerator = colSums(samples)
+#'  denominator = apply( samples, MARGIN = 2, FUN = \(x) weighted.mean(x=x, w=x) )
+#'  c(
+#'    Estimate = mean( numerator / denominator ),
+#'    SD = sd( numerator / denominator ),
+#'    True = sum(x_s) / weighted.mean(x_s, w = x_s)
+#'  )
+#'
 #' @export
 integrate_output <-
 function( object,
@@ -376,6 +435,12 @@ function( object,
     out = data.frame( "Estimate"=rep$Metric, "Std. Error"=NA, "Est. (bias.correct)"=NA, "Std. (bias.correct)"=NA )
   }
   
+  # Replace missing levels of block with NAs
+  which_missing = which(table(factor(block,levels = seq_len(max(block)))) == 0)
+  if(length(which_missing)>0){
+    out[which_missing,] = NA
+  }
+
   # deal with memory internally
   remove("newobj")
   gc() 
@@ -494,35 +559,57 @@ function( object,
   W2_gl = make_spatial_varying( object$internal$delta_spatial_varying )
 
   # Assemble A_gs
-  if( is(object$spatial_domain, "fm_mesh_2d") ){
+  if( is(object$spatial_domain,"vertex_coords") ){
+    A_gs = fm_evaluator( object$spatial_domain, loc=as.matrix(newdata[,object$internal$space_columns]) )$proj$A
+  }else if( is(object$spatial_domain, "fm_mesh_2d") ){
     A_gs = fm_evaluator( object$spatial_domain, loc=as.matrix(newdata[,object$internal$space_columns]) )$proj$A
   }else if( is(object$spatial_domain, "igraph") ) {
     Match = match( newdata[,object$internal$space_columns], rownames(object$tmb_inputs$tmb_data$Adj) )
-    if(any(is.na(Match))) stop("Check `spatial_domain` for SAR")
+    if(any(is.na(Match))) stop("Check `object$spatial_domain` for SAR")
     A_gs = sparseMatrix( i=seq_len(nrow(newdata)), j=Match, x=rep(1,nrow(newdata)) )
   }else if( is(object$spatial_domain,"sfnetwork_mesh") ){      # if( !is.null(sem) )
     # stream network
     A_gs = sfnetwork_evaluator( stream = object$spatial_domain$stream,
                                 loc = as.matrix(newdata[,object$internal$space_columns]) )
-  }else if( is(object$spatial_domain, "sfc_GEOMETRY") ){
+  }else if( is_areal_sf(object$spatial_domain) ){
     sf_coords = st_as_sf( newdata,
                           coords = object$internal$space_columns,
                           crs = st_crs(object$spatial_domain) )
-    s_i = as.integer(st_within( sf_coords, object$spatial_domain ))
-    if(any(is.na(s_i))){
+    s_g = as.integer(st_within( sf_coords, object$spatial_domain ))
+    if(any(is.na(s_g))){
       stop("Some rows of `newdata` in `predict` are not within the SAR domain.
             Please exclude rows listed below or refit model with extended domain.\n",
-            paste0(which(is.na(s_i)), collapse = ", ") )
+            paste0(which(is.na(s_g)), collapse = ", ") )
     }
-    A_gs = sparseMatrix( i = seq_along(s_i),
-                         j = s_i,
+    A_gs = sparseMatrix( i = seq_along(s_g),
+                         j = s_g,
                          x = 1,
-                         dims = c(length(s_i),length(object$spatial_domain)) )
-  }else{
+                         dims = c(length(s_g),length(object$spatial_domain)) )
+  }else if( is(object$spatial_domain,"nngp_domain") ){
+    sf_coords = st_as_sf( newdata,
+                          coords = object$internal$space_columns,
+                          crs = st_crs(object$spatial_domain$sf_areal) )
+    s_g = as.integer(st_within( sf_coords, object$spatial_domain$sf_areal ))
+    if(any(is.na(s_g))){
+      stop("Some rows of `newdata` in `predict` are not within the NNGP domain.
+            Please exclude rows listed below or refit model with extended domain.\n",
+            paste0(which(is.na(s_g)), collapse = ", ") )
+    }
+    A_gs = sparseMatrix( i = seq_along(s_g),
+                         j = s_g,
+                         x = 1,
+                         dims = c(length(s_g),length(object$spatial_domain$sf_areal)) )
+  }else if( is.null(object$spatial_domain) ) {
     A_gs = matrix(1, nrow=nrow(newdata), ncol=1)    # dgCMatrix
     A_gs = as(Matrix(A_gs),"CsparseMatrix")
+  }else{
+    stop("`spatial_domain` is does not match options:  class fm_mesh_2d, igraph, nngp_domain, sfnetwork_mesh, sfc_GEOMETRY, or NULL")
   }
+
   predAtriplet = Matrix::mat2triplet(A_gs)
+  if( isFALSE(all.equal(rowSums(A_gs), 1, tol = 0.01)) ){
+    warning("Some predictions appear to be outside spatial domain" )
+  }
 
   # Turn of t_i and c_i when times and variables are missing, so that delta_k isn't built
   if( length(object$internal$times) > 0 ){